Rename theta packages, too
