+++ src/main/scala/org/apache/spark/sql/datasketches/theta/aggregate/ThetaSketchAggUnion.scala
@@ -15,16 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.aggregate
+package org.apache.spark.sql.datasketches.theta.aggregate
 
 import org.apache.datasketches.memory.Memory
 import org.apache.datasketches.theta.{Sketch, SetOperation}
+import org.apache.spark.sql.datasketches.theta.types.{ThetaSketchType, ThetaSketchWrapper}
 
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.{ExpectsInputTypes, Expression, ExpressionDescription, Literal}
 import org.apache.spark.sql.catalyst.expressions.aggregate.TypedImperativeAggregate
 import org.apache.spark.sql.catalyst.trees.BinaryLike
-import org.apache.spark.sql.types.{AbstractDataType, DataType, IntegerType, ThetaSketchWrapper, ThetaSketchType}
+import org.apache.spark.sql.types.{AbstractDataType, DataType, IntegerType}
 
 /**
  * Theta Union operation.
